Claude Apps/python-ai-roadmap/phase2/module-l/pyproject.toml
[project]
name = "module-l"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.2",
    "torchvision>=0.17",
    "streamlit>=1.35",
    "streamlit-drawable-canvas>=0.9.3",
    "numpy>=1.26",
    "pillow>=10.0",
    "pandas>=2.0",
]

[dependency-groups]
dev = ["pytest>=8.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
